requirements.txt
numpy>=1.20.0
torch>=1.9.0
transformers>=4.18.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
pandas>=1.3.0
tqdm>=4.62.0
openai>=1.0.0
python-dotenv>=0.19.0
pytest>=6.2.5
